refactor docs and add initial Jest testing
