--- src/docgen/genDocs.ts
@@ -1,33 +1,0 @@
-import DocsBuilder from "./DocsBuilder.js";
-import path from "path";
-import DocUtil from "../utils/DocsUtil.js";
-import ProjectUtil from "../utils/ProjectUtil.js";
-import Config, { CONFIG_NAME } from "../config/Config.js";
-
-async function main() {
-  try {
-    const projectRoot = ProjectUtil.findProjectRoot();
-    if (!projectRoot) {
-      throw new Error(
-        "Project root not found. Make sure you are running this from within a Node.js project.",
-      );
-    }
-
-    const config = new Config();
-    config.load();
-
-    const baseOutputDir = config.docs.outputDir || "bin";
-    const bugpromptDir = CONFIG_NAME;
-    const fullOutputPath = path.join(projectRoot, baseOutputDir, bugpromptDir);
-
-    // Ensure output dir exists and is clean
-    await DocUtil.prepareOutputDirectory(fullOutputPath);
-
-    const builder = new DocsBuilder(config.docs, fullOutputPath);
-    await builder.build();
-  } catch (error) {
-    console.error("Error in documentation generation:", error);
-  }
-}
-
-main();
